Use framework resources for paste menuId
For paste menu IDs we override the given title, content decription,
icon, and tooltip text  with the framework's resources for those
actions. This way apps are unable to trick the user as somethign else.

We also apply localization correctly, even though Configuration is sent
and the LocaleList is identical the system drawn, even when inflated
with the new configuration context,  TextView somehow ignores any app
specific locale and falls back to the sysetm configuredlocales. This
might be a TextView bug.

Test: Manual
Bug: 363318732
Fixes: 404562081
Flag: android.permission.flags.system_selection_toolbar_enabled

Change-Id: I8b24d21d1ca2089604586a5836fd134616c0f6e1