Switch exceptions to logs for client controlled bubble requests
If there is an issue with a client request, because the verification
of the params and the state happens on the main executor, an
exception would crash WM Shell instead of being propagated back to
the caller. This is not helpful if the issue is with the client's
request.

Instead switch to logging the error and plan to add a dedicated
callback to inform the clients when the bubbles they requested are
removed by the user.

Bug: 407149510
Test: Use DroidNavigator app and manually remove the bubble
Flag: EXEMPT bugfix
Change-Id: Ib3d5571470251bbe09351e1aa02092b18fc93979