Only check for xml flag attributes when we know there is one there
This makes it so that when we are flattening xml files we figure out
which elements have the featureFlag attribute and persist that in the
binary xml. Then at runtime we can quickly check that before we try and
find the actual attribute.

Test: Automation
Flag: android.content.res.layout_readwrite_flags
Bug: 398541237
Change-Id: I1015dd45c4a0ba38bea75cad30b59cec4a05d825