Clean up bubbles expanded state in sysui state
When the last bubble is removed and we are processing the update, make
sure we update sysui state that bubbles are no longer expanded.
BubbleStackView and BubbleBarLayerView should be updating the state.
But in case of bugs like b/391909607 where the state gets corrupted, we
don't want to leave the device in a state where the home gesture stops
working.
If we leave the sysui state as bubbles expanded, the home gesture is
ignored.

Bug: 391909607
Flag: com.android.wm.shell.enable_bubble_swipe_up_cleanup
Test: manual, exact repro steps in the linked bug
Change-Id: Ic7334afb1d5e5460bac3a906a8590d28bb5502d4