Flag InsetsSource as invalid if mServerVisible is false
Since this CL, if an insets source is invalid, its visibility cannot be
overridden, as if the source is null.

This CL also enables the caller of InsetsState#equals to skip invalid
sources so that they can avoid the redundant consequence triggered by
adding an invalid insets source.

Fix: 406465708
Flag: EXEMPT bugfix
Test: atest InsetsStateTest InsetsStateControllerTest
      WindowInsetsControllerTests
Change-Id: I42a42ce414d4449bf1e89b95bd8a710ae4a157ec