Move listener additions within AccessibilityRepository to background
We should specify our own background Handlers and Executors
so that the binder calls don't happen on the main thread.

The overloaded signature without these default to the main
thread, without being obvious they are doing so.

The tests also now force a non-nulls to be passed too.

Fixes: 409473208
Test: atest AccessibilityRepositoryTest.kt
Flag: EXEMPT bugfix
Change-Id: Ie22cf8877bf585d1a7d0ebf28f368c288024a4e8