Add "Include default display in topology" switch for projected mode
If the device is not in projected mode or if the switch is enabled, the
default display is always included in the topology. Otherwise:

- when an external display is added, the default display will be removed
  from the topology;
- when all the external displays are removed, the default display will
  be added back to the topology;
- when the user enables/disables the switch, the default display will be
  added to / removed from the topology.

This switch is only available for the device in projected mode, and when
the "Mirror built-in display" switch is disabled.

Bug: 407512389
Test: adb shell settings put secure include_default_display_in_topology
{1|0}
Test: atest DisplayManagerServiceTest
Flag: com.android.server.display.feature.flags.enable_default_display_in_topology_switch

Change-Id: I3ee34e3330e4c99a4e6cee2ddd45631b47d2d4b7