Fix crash due to missing OnBackPressedDispatcherOwner
LocalOnBackPressedDispatcherOwner was not set for flexiglass off,
therefore the device would get into a crashloop when a BundleHeader
was initialized. We set it locally since setting it in
KeyguardRootViewBinder did not work. Though there might be a higher
level View where we could set this once instead of for each
ComposeView. Other ComposeView like the QsHeader do set it locally as
well.

Bug: 411527659
Test: reprod crash without flexiglass and verified fix
Flag: com.android.systemui.notification_bundle_ui
Change-Id: Iaaab3ee2e3a7cd493d15a73fb238e5365349872b