Add security info to dual shade toolbar
This adds the security info in the toolbar. The security info shows icon
and text for 5s whenever the shade is open or when the value changes (to
non-empty). After that, it collapses into just the icon. When it's
expanded with text, only the power button remains.

This does not implement final UI (paddings).

Test: manual, turn on/off vpn
Test: atest ToolbarViewModelTest
Flag: com.android.systemui.scene_container
Bug: 415030295

Change-Id: I799fbbb796855e5a55b6461455c98974f94e021c