Extending the number of Roundable targets during magnetic pulls.
To avoid sharp corners while pulling magnetic notifications, we are
extending the set of identified roundable targets when swiping a
magnetic notification. This is possible by caching roundable targets as
a list in the NotificationRoundnessManager(as opposed to only 3), and
extending the algorithm in NotificationTargetsHelper to find both
magnetic and roundable targets when swiping a notification. For a fixed
number of targets, we also extend the result of the finder algorithm so
that two additional roundable boundaries are included.

These changes ensure that notifications don't have sharp corners during
a magnetic swipe.

Test: modified Unit tests
Test: manual. Verified corners are rounded on all notifications as a
  magnetic notification is pulled. This was verified for notifications
  in a group and in the stack.
Flag: com.android.systemui.magnetic_notification_swipes
Bug: 413925464
Change-Id: I1a9e30cb7ef7fc307f7f2b4efadd179f51afc2de