Fix groups within bundles expanding
Before this change mChildrenExpanded would be set to true for all
children when expanding a bundle. It assumed there would be no groups
within groups. This breaks the initial ui state of a group within a
bundle (it would correct itself when expanding/collapsing a group once)

This change makes sure that expansion is still propagated
to child notifications but not to groups.

Bug: b/389839492
Test: manual tested expanding bundle and verifying group is displayed
correctly
Flag: com.android.systemui.notification_bundle_ui

Change-Id: I978a73b1cf7f2c44f60c94de18088939a30b89a1