[PiP on Desktop] Always expand PiP to fullscreen if mid-recents.
This is a workaround for an issue where expanding PiP while in Recents
during Desktop session will lead to the device getting stuck in a bad
state. This is a temporary fix by forcing PiP to expand to fullscreen
always, if we are expanding while in Recents in a Desktop session.

Bug: 409201669
Test: atest PipDesktopStateTest
Flag: com.android.window.flags.enable_desktop_windowing_pip

Change-Id: Ibcdc47054aea52092341e55e2ee435543db58d5d