Desks: Handle back-nav transitions correctly
When a task is moved to back using back navigation, handle the
transition request and make sure to minimize it correctly using
DesksOrganizer (so that it is reparented to the minimization root). Also
stop triggering desks deactivations in handleFreeformTaskLaunch when the
desk is already inactive.

In the DesktopBackNavTransitionObserver, do not assume that non-freeform
tasks are not desktop tasks. Instead, use DesksOrganizer or
DesktopRepository to check for whether a task is in desktop or part
exiting a desk.

The above changes allows back navigation of tasks to be correctly
tracked as minimizations in both the repository and for animation
purposes.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Fix: 413024179
Fix: 413135711
Fix: 413134745
Test: back nav the last window or a task when it is not the last window
and verify animations work as expected without flickering. Also verify
minimize indicator is present after back nav and the app can be reopened
from taskbar.

Change-Id: Ifd2791464bbf8adccf872d11627e78a9ffcfc13d