Allow TaskView to move windows
Adds the ability to mark that a task view is moving windows. When
marked, visibility changes and attach detach events are ignored
in order to preserve the existing surface. This allows users of
TaskView to reparent the surface into the new window as needed.

Bug: 392893178
Flag: EXEMPT new API no-op change
Test: EXEMPT new API that isn't currently triggered

Change-Id: I3ecebccafec9acdfd27feac014eac44cefe64c18