Fade out behind scrim when transitioning from lockscreen to dream
The lockscreen has a scrim underneath it to dim the wallpaper. This
layer is on top of the dream, so transitions from the lock screen to
dream end with the dream abruptly getting lighter when the sysui window
is no longer visible.

This change fades the behind scrim out as the dream starts. The
alternate bouncer is not covered here as it needs more work.

Bug: 409720192
Test: atest PrimaryBouncerToDreamingTransitionViewModelTest LockscreenToDreamingTransitionViewModelTest
      (see bug for video)
Flag: com.android.systemui.dream_transition_fixes
Change-Id: I393bf308b25a63bab9374003e3fcf00e4461d2d4